Use state file for application persistence
