+++ pt_os_web_portal/backend/__init__.py
@@ -5,7 +5,7 @@ from flask_sockets import Sockets
 sockets: Sockets
 
 
-def create_app(test_mode, os_updater):
+def create_app(test_mode, os_updater, state_manager):
     app = Flask(
         __name__, static_url_path="", static_folder="./build", template_folder="./build"
     )
@@ -17,6 +17,8 @@ def create_app(test_mode, os_updater):
         app.config["TESTING"] = True
     if os_updater:
         app.config["OS_UPDATER"] = os_updater
+    if state_manager:
+        app.config["STATE_MANAGER"] = state_manager
 
     with app.app_context():
         from . import routes
