+++ pt_os_web_portal/backend/helpers/finalise.py
@@ -1,10 +1,10 @@
 from os import path, remove
-from pathlib import Path
 
 from pitop.common.command_runner import run_command, run_command_background
 from pitop.common.current_session_info import get_user_using_display
 from pitop.common.logger import PTLogger
 
+from ..routes import get_state_manager
 from .paths import use_test_path
 
 
@@ -46,6 +46,7 @@ def deprioritise_openbox_session() -> None:
 def stop_onboarding_autostart() -> None:
     PTLogger.debug("Function: stop_onboarding_autostart()")
     remove("/etc/xdg/autostart/pt-os-setup.desktop")
+    get_state_manager().set("app", "state", "onboarding")
 
 
 def enable_firmware_updater_service():
@@ -108,11 +109,7 @@ def python_sdk_docs_url():
 
 
 def onboarding_completed():
-    PTLogger.debug("Function: onboarding_completed()")
-    try:
-        return not Path("/etc/xdg/autostart/pt-os-setup.desktop").exists()
-    except Exception:
-        return False
+    return get_state_manager().get("app", "state") != "onboarding"
 
 
 def open_further():
